Fix skinned results/song select grades not displaying

+++ Quaver.Shared/Screens/Results/UI/Header/Contents/ResultsScreenHeaderContentContainer.cs
@@ -187,8 +187,8 @@ namespace Quaver.Shared.Screens.Results.UI.Header.Contents
             GradeSprite.Image = SkinManager.Skin.GradesLarge[grade];
 
             const int width = 110;
-
-            GradeSprite.Size = new ScalableVector2(width, GradeSprite.Image.Height / GradeSprite.Image.Width * width);
+            
+            GradeSprite.Size = new ScalableVector2(width, (float) GradeSprite.Image.Height / GradeSprite.Image.Width * width);
             GradeSprite.Y = -TabSelector.Height - 22;
         }
